Crate turbo_tasks

A task scheduling and caching system that is focused on incremental execution.

It defines 4 primitives:

functions: Unit of execution, invalidation and reexecution.

values: Data created, stored and returned by functions.

traits: Traits that define a set of functions on values.

collectibles: Values emitted in functions that bubble up the call graph and can be collected in parent functions.

It also defines some derived elements from that:

cells: The locations in functions where values are stored. The content of a cell can change after the reexecution of a function.

Vcs: A reference to a cell in a function or a return value of a function.

task: An instance of a function together with its arguments.

A Vc can be read to get a read-only reference to the stored data.

On execution of functions, turbo-tasks will track which Vcs are read. Once any of these change, turbo-tasks will invalidate the task created from the function’s execution and it will eventually be scheduled and reexecuted.

Collectibles go through a similar process.

Modules

backend

debug

event

graph

persisted_graph

primitives

registry

small_duration

test_helpers

trace

util

Structs

CellId

CollectiblesFuture

Completion

Just an empty type, but it’s never equal to itself. CompletionVc can be used as return value instead of () to have a concrete reference that can be awaited. It will invalidate the awaiting task everytime the referenced task has been executed.

CompletionVc

Vc for Completion

CompletionsVc

Vc for [Completions]

Error

The Error type, a wrapper around a dynamic error type.

FunctionId

InvalidationReasonSet

A set of InvalidationReasons. They are automatically deduplicated and merged by kind during insertion. It implements Display to get a readable representation.

Invalidator

NativeFunction

A native (rust) turbo-tasks function. It’s used internally by #[turbo_tasks::function].

NativeFunctionVc

Vc for NativeFunction

Nothing

Just an empty type. NothingVc can be used as return value instead of () to have a concrete reference that can be awaited.

NothingVc

Vc for Nothing

ReadRawVcFuture

ReadRef

The read value of a value cell. The read value is immutable, while the cell itself might change over time. It’s basically a snapshot of a value at a certain point in time.

SharedReference

SharedValue

State

TaskId

TraitCast

Casts an arbitrary cell content into a TraitRef<T>.

TraitRef

Similar to a [ReadRef<T>], but contains a value trait object instead. The only way to interact with a TraitRef<T> is by passing it around or turning it back into a value trait vc by calling [ReadRef::cell].

TraitType

TraitTypeId

TransientInstance

Pass a reference to an instance to a turbo-tasks function.

TransientValue

Pass a value by value (Value<Xxx>) instead of by reference (XxxVc).

TransparentValueCast

Casts an arbitrary cell content into a ReadRef<T, U>.

TurboTasks

Unused

A wrapper around a value that is unused.

UpdateInfo

Value

Pass a value by value (Value<Xxx>) instead of by reference (XxxVc).

ValueCast

Casts an arbitrary cell content into a ReadRef<T, U>.

ValueToStringVc

ValueType

A definition of a type of data.

ValueTypeId

Enums

RawVc

ResolveTypeError

StatsType

The type of stats reporting.

TaskInput

Traits

CollectiblesSource

DynamicEqHash

FromSubTrait

FromTaskInput

IdMapping

IntoSuperTrait

IntoTraitRef

A trait that allows a value trait vc to be converted into a trait reference.

InvalidationReason

A user-facing reason why a task was invalidated. This should only be used for invalidation that were triggered by the user.

InvalidationReasonKind

Invalidation reason kind. This is used to merge multiple reasons of the same kind into a combined description.

JoinIterExt

TaskIdProvider

TraitMethod

TryJoinIterExt

TurboTasksApi

TurboTasksBackendApi

TurboTasksCallApi

Typed

TypedForInput

Marker trait that a turbo_tasks::value is prepared for serialization as Value<…> input. Either use #[turbo_tasks::value(serialization: auto_for_input)] or avoid Value<…> in favor of a real Vc

ValueToString

ValueTraitVc

ValueVc

Functions

dynamic_call

see TurboTasks dynamic_call

emit

get_invalidator

Get an Invalidator that can be used to invalidate the current [Task] based on external events.

mark_finished

Marks the current task as finished. This excludes it from waiting for strongly consistency.

mark_stateful

Marks the current task as stateful. This prevents the tasks from being dropped without persisting the state.

register

run_once

run_once_with_reason

spawn_blocking

spawn_thread

trait_call

see TurboTasks trait_call

turbo_tasks

with_task_id_mapping

without_task_id_mapping

Type Definitions

Result

Result<T, Error>

Attribute Macros

function

value

Creates a ValueVc struct for a struct or enum that represent that type placed into a cell in a Task.

value_impl

value_trait