Enum turbo_tasks::RawVc

pub enum RawVc {
    TaskOutput(TaskId),
    TaskCell(TaskId, CellId),
}

Variants

TaskOutput(TaskId)

TaskCell(TaskId, CellId)

Implementations

impl RawVc

pub fn into_read<T: Any + Send + Sync>(self) -> ReadRawVcFuture<ValueCast<T>>

pub fn into_strongly_consistent_read<T: Any + Send + Sync>( self ) -> ReadRawVcFuture<ValueCast<T>>

pub unsafe fn into_transparent_read<T: Any + Send + Sync, U: Any + Send + Sync>( self ) -> ReadRawVcFuture<TransparentValueCast<T, U>>

Safety

T and U must be binary identical (#[repr(transparent)])

pub unsafe fn into_transparent_strongly_consistent_read<T: Any + Send + Sync, U: Any + Send + Sync>( self ) -> ReadRawVcFuture<TransparentValueCast<T, U>>

Safety

T and U must be binary identical (#[repr(transparent)])

pub fn into_trait_read<T>(self) -> ReadRawVcFuture<TraitCast<T>> where T: ValueTraitVc,

pub fn into_strongly_consistent_trait_read<T>( self ) -> ReadRawVcFuture<TraitCast<T>> where T: ValueTraitVc,

pub async fn into_read_untracked<T: Any + Send + Sync>( self, turbo_tasks: &dyn TurboTasksApi ) -> Result<ReadRef<T>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

pub async fn into_strongly_consistent_read_untracked<T: Any + Send + Sync>( self, turbo_tasks: &dyn TurboTasksApi ) -> Result<ReadRef<T>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

pub async fn internal_pointer_untracked( self, turbo_tasks: &dyn TurboTasksApi ) -> Result<SharedReference>

Returns the hash of the pointer that holds the Vc’s current data. This value will change every time the TaskCell recomputes.

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

pub async fn resolve_trait( self, trait_type: TraitTypeId ) -> Result<Option<RawVc>, ResolveTypeError>

pub async fn resolve_value( self, value_type: ValueTypeId ) -> Result<Option<RawVc>, ResolveTypeError>

pub async fn resolve(self) -> Result<RawVc>

Resolve the reference until it points to a cell directly.

Resolving will wait for task execution to be finished, so that the returned Vc points to a cell that stores a value.

Resolving is necessary to compare identities of Vcs.

This is async and will rethrow any fatal error that happened during task execution.

pub async fn resolve_strongly_consistent(self) -> Result<RawVc>

Resolve the reference until it points to a cell directly in a strongly consistent way.

Resolving will wait for task execution to be finished, so that the returned Vc points to a cell that stores a value.

Resolving is necessary to compare identities of Vcs.

This is async and will rethrow any fatal error that happened during task execution.

pub fn connect(&self)

pub fn is_resolved(&self) -> bool

pub fn get_task_id(&self) -> TaskId