Module turbo_tasks::graph

Structs

GraphTraversal

GraphTraversal is a utility type that can be used to traverse a graph of nodes, where each node can have a variable number of outgoing edges. The traversal is done in parallel, and the order of the nodes in the traversal result is determined by the GraphStore parameter.

NonDeterministic

A graph traversal that does not guarantee any particular order, and may not return the same order every time it is run.

ReverseTopological

A graph traversal that returns nodes in reverse topological order.

SkipDuplicates

A GraphStore wrapper that skips nodes that have already been visited. This is necessary to avoid repeated work when traversing non-tree graphs (i.e. where a node can have more than one incoming edge).

Enums

GraphTraversalResult

VisitControlFlow

The control flow of visiting an edge during a graph traversal.

Traits

GraphStore

A graph store is a data structure that will be built up during a graph traversal. It is used to store the results of the traversal.

Visit

A trait that allows a graph traversal to visit the edges of a node transitively.