Struct turbo_tasks::graph::AdjacencyMap

pub struct AdjacencyMap<T>where
    T: Eq + Hash + Clone,{ /* private fields */ }

A graph traversal that builds an adjacency map

Implementations

impl<T> AdjacencyMap<T>where T: Eq + Hash + Clone,

pub fn new() -> Self

Creates a new adjacency map

pub fn roots(&self) -> impl Iterator<Item = &T>

Returns an iterator over the root nodes of the graph

pub fn get(&self, node: &T) -> Option<impl Iterator<Item = &T>>

Returns an iterator over the children of the given node

impl<T> AdjacencyMap<T>where T: Eq + Hash + Clone,

pub fn into_reverse_topological(self) -> ReverseTopologicalIter<T>

Returns an iterator over the nodes in reverse topological order, starting from the roots.

pub fn reverse_topological_from_node<'graph>( &'graph self, node: &'graph T ) -> ReverseTopologicalFromNodeIter<'graph, T>

Returns an iterator over the nodes in reverse topological order, starting from the given node.