Trait turbo_tasks::graph::GraphTraversal

pub trait GraphTraversal: GraphStore + Sized {
    // Required methods
    fn visit<RootEdgesIt, VisitImpl, Abort, Impl>(
        self,
        root_edges: RootEdgesIt,
        visit: VisitImpl
    ) -> GraphTraversalFuture<Self, VisitImpl, Abort, Impl>
       where VisitImpl: Visit<Self::Node, Abort, Impl>,
             RootEdgesIt: IntoIterator<Item = VisitImpl::Edge>;
    fn skip_duplicates(self) -> SkipDuplicates<Self>;
}

GraphTraversal is a utility type that can be used to traverse a graph of nodes, where each node can have a variable number of outgoing edges. The traversal is done in parallel, and the order of the nodes in the traversal result is determined by the GraphStore parameter.

Required Methods

fn visit<RootEdgesIt, VisitImpl, Abort, Impl>( self, root_edges: RootEdgesIt, visit: VisitImpl ) -> GraphTraversalFuture<Self, VisitImpl, Abort, Impl>where VisitImpl: Visit<Self::Node, Abort, Impl>, RootEdgesIt: IntoIterator<Item = VisitImpl::Edge>,

fn skip_duplicates(self) -> SkipDuplicates<Self>