Struct turbo_tasks::primitives::Bytes

#[repr(transparent)]
pub struct Bytes(_);

Implementations

impl Bytes

pub fn cell(self) -> BytesVc

Places a value in a cell of the current task.

Cell is selected based on the value type and call order of cell.