Struct turbo_tasks::ReadRef

pub struct ReadRef<T, U = T>(_, _);

The read value of a value cell. The read value is immutable, while the cell itself might change over time. It’s basically a snapshot of a value at a certain point in time.

Internally it stores a reference counted reference to a value on the heap.

Invariant: T and U are binary identical (#[repr(transparent)])

Implementations

impl<T> ReadRef<T, T>

pub fn new(arc: Arc<T>) -> Self

impl<T, U> ReadRef<T, U>

pub unsafe fn new_transparent(arc: Arc<T>) -> Self

Safety

T and U must be binary identical (#[repr(transparent)])

pub fn ptr_eq(&self, other: &ReadRef<T, U>) -> bool

pub fn ptr_cmp(&self, other: &ReadRef<T, U>) -> Ordering

impl<T, U> ReadRef<T, U>where T: Typed + Send + Sync + 'static,

pub fn cell(trait_ref: ReadRef<T, U>) -> T::Vc

Returns a new cell that points to a value that implements the value trait T.

impl<T, U: Clone> ReadRef<T, U>

pub fn clone_value(&self) -> U

This will clone the contained value instead of cloning the ReadRef. This clone is more expensive, but allows to get an mutable owned value.