Struct turbo_tasks::TurboTasks

pub struct TurboTasks<B: Backend + 'static> { /* private fields */ }

Implementations

impl<B: Backend + 'static> TurboTasks<B>

pub fn new(backend: B) -> Arc<Self>

pub fn pin(&self) -> Arc<Self>

pub fn spawn_root_task( &self, functor: impl Fn() -> Pin<Box<dyn Future<Output = Result<RawVc>> + Send>> + Sync + Send + 'static ) -> TaskId

Creates a new root task

pub fn spawn_once_task( &self, future: impl Future<Output = Result<RawVc>> + Send + 'static ) -> TaskId

Creates a new root task, that is only executed once. Dependencies will not invalidate the task.

pub async fn run_once<T: TraceRawVcs + Send + 'static>( &self, future: impl Future<Output = Result<T>> + Send + 'static ) -> Result<T>

pub fn dynamic_call(&self, func: FunctionId, inputs: Vec<TaskInput>) -> RawVc

Calls a native function with arguments. Resolves arguments when needed with a wrapper [Task].

pub fn trait_call( &self, trait_type: TraitTypeId, trait_fn_name: Cow<'static, str>, inputs: Vec<TaskInput> ) -> RawVc

Calls a trait method with arguments. First input is the self object. Uses a wrapper task to resolve

pub async fn wait_foreground_done(&self)

pub fn get_in_progress_count(&self) -> usize

pub async fn wait_task_completion( &self, id: TaskId, fully_settled: bool ) -> Result<()>

pub async fn get_or_wait_update_info( &self, aggregation: Duration ) -> (Duration, usize)

๐Deprecated: Use get_or_wait_aggregated_update_info instead

pub async fn update_info( &self, aggregation: Duration, timeout: Duration ) -> Option<(Duration, usize)>

๐Deprecated: Use aggregated_update_info instead

pub async fn get_or_wait_aggregated_update_info( &self, aggregation: Duration ) -> UpdateInfo

Returns UpdateInfo with all updates aggregated over a given duration (aggregation). Will wait until an update happens.

pub async fn aggregated_update_info( &self, aggregation: Duration, timeout: Duration ) -> Option<UpdateInfo>

Returns UpdateInfo with all updates aggregated over a given duration (aggregation). Will only return None when the timeout is reached while waiting for the first update.

pub async fn wait_background_done(&self)

pub async fn stop_and_wait(&self)

pub fn backend(&self) -> &B