Trait turbo_tasks::trace::TraceRawVcs

pub trait TraceRawVcs {
    // Required method
    fn trace_raw_vcs(&self, context: &mut TraceRawVcsContext);

    // Provided method
    fn get_raw_vcs(&self) -> Vec<RawVc>  { ... }
}

Trait that allows to walk data to find all RawVcs contained.

This is important for Garbage Collection to mark all Cells and therefore Tasks that are still in use.

It can also be used to optimize transferring of Tasks, where knowning the referenced Cells/Tasks allows pushing them earlier.

#[derive(TraceRawVcs)] is available. #[trace_ignore] can be used on fields to skip tracing for them.

Required Methods

fn trace_raw_vcs(&self, context: &mut TraceRawVcsContext)

Provided Methods

fn get_raw_vcs(&self) -> Vec<RawVc>