Trait turbo_tasks::IntoTraitRef

pub trait IntoTraitRef {
    type TraitVc: ValueTraitVc;
    type Future: Future<Output = Result<TraitRef<Self::TraitVc>>>;

    fn into_trait_ref(self) -> Self::Future;
}

A trait that allows a value trait vc to be converted into a trait reference.

The signature is similar to IntoFuture, but we don’t want trait vcs to have the same future-like semantics as value vcs when it comes to producing refs. This behavior is rarely needed, so in most cases, .awaiting a trait vc is a mistake.

Required Associated Types

type TraitVc: ValueTraitVc

type Future: Future<Output = Result<TraitRef<Self::TraitVc>>>

Required Methods

fn into_trait_ref(self) -> Self::Future