Trait turbo_tasks::TurboTasksApi

pub trait TurboTasksApi: TurboTasksCallApi + Sync + Send {

    fn invalidate(&self, task: TaskId);
    fn notify_scheduled_tasks(&self);
    fn try_read_task_output(
        &self,
        task: TaskId,
        strongly_consistent: bool
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_output_untracked(
        &self,
        task: TaskId,
        strongly_consistent: bool
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_cell(
        &self,
        task: TaskId,
        index: CellId
    ) -> Result<Result<CellContent, EventListener>>;
    fn try_read_task_cell_untracked(
        &self,
        task: TaskId,
        index: CellId
    ) -> Result<Result<CellContent, EventListener>>;
    fn try_read_task_collectibles(
        &self,
        task: TaskId,
        trait_id: TraitTypeId
    ) -> Result<Result<AutoSet<RawVc>, EventListener>>;
    fn emit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc);
    fn unemit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc);
    fn unemit_collectibles(
        &self,
        trait_type: TraitTypeId,
        collectibles: &AutoSet<RawVc>
    );
    fn try_read_own_task_cell_untracked(
        &self,
        current_task: TaskId,
        index: CellId
    ) -> Result<CellContent>;
    fn read_current_task_cell(&self, index: CellId) -> Result<CellContent>;
    fn update_current_task_cell(&self, index: CellId, content: CellContent);

}

Required Methods

fn invalidate(&self, task: TaskId)

fn notify_scheduled_tasks(&self)

Eagerly notifies all tasks that were scheduled for notifications via schedule_notify_tasks_set()

fn try_read_task_output(
&self,
task: TaskId,
strongly_consistent: bool
) -> Result<Result<RawVc, EventListener>>

fn try_read_task_output_untracked(
&self,
task: TaskId,
strongly_consistent: bool
) -> Result<Result<RawVc, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn try_read_task_cell(
&self,
task: TaskId,
index: CellId
) -> Result<Result<CellContent, EventListener>>

fn try_read_task_cell_untracked(
&self,
task: TaskId,
index: CellId
) -> Result<Result<CellContent, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn try_read_task_collectibles(
&self,
task: TaskId,
trait_id: TraitTypeId
) -> Result<Result<AutoSet<RawVc>, EventListener>>

fn emit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc)

fn unemit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc)

fn unemit_collectibles(
&self,
trait_type: TraitTypeId,
collectibles: &AutoSet<RawVc>
)

fn try_read_own_task_cell_untracked(
&self,
current_task: TaskId,
index: CellId
) -> Result<CellContent>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn read_current_task_cell(&self, index: CellId) -> Result<CellContent>

fn update_current_task_cell(&self, index: CellId, content: CellContent)