Attribute Macro turbo_tasks::value

#[value]

Creates a ValueVc struct for a struct or enum that represent that type placed into a cell in a Task.

That ValueVc object can be .await?ed to get a readonly reference to the original value.

into argument (#[turbo_tasks::value(into: xxx)])

When provided the ValueVc implement From<Value> to allow to convert a Value to a ValueVc by placing it into a cell in a Task.

into: new: Always overrides the value in the cell. Invalidating all dependent tasks.

into: shared: Compares with the existing value in the cell, before overriding it. Requires Value to implement [Eq].

TODO: add more documentation: presets, traits