Trait turbo_tasks::backend::Backend

pub trait Backend: Sync + Send {
    type ExecutionScopeFuture<T: Future<Output = Result<()>> + Send + 'static>: Future<Output = Result<()>> + Send + 'static;

    fn invalidate_task(
        &self,
        task: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn invalidate_tasks(
        &self,
        tasks: Vec<TaskId>,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn get_task_description(&self, task: TaskId) -> String;
    fn execution_scope<T: Future<Output = Result<()>> + Send + 'static>(
        &self,
        task: TaskId,
        future: T
    ) -> Self::ExecutionScopeFuture<T>;
    fn try_start_task_execution(
        &self,
        task: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Option<TaskExecutionSpec>;
    fn task_execution_result(
        &self,
        task: TaskId,
        result: Result<Result<RawVc>, Option<Cow<'static, str>>>,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn task_execution_completed(
        &self,
        task: TaskId,
        duration: Duration,
        instant: Instant,
        stateful: bool,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> bool;
    fn run_backend_job<'a>(
        &'a self,
        id: BackendJobId,
        turbo_tasks: &'a dyn TurboTasksBackendApi
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>>;
    fn try_read_task_output(
        &self,
        task: TaskId,
        reader: TaskId,
        strongly_consistent: bool,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_output_untracked(
        &self,
        task: TaskId,
        strongly_consistent: bool,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_cell(
        &self,
        task: TaskId,
        index: CellId,
        reader: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<Result<CellContent, EventListener>>;
    fn try_read_task_cell_untracked(
        &self,
        task: TaskId,
        index: CellId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<Result<CellContent, EventListener>>;
    fn try_read_task_collectibles(
        &self,
        task: TaskId,
        trait_id: TraitTypeId,
        reader: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<Result<AutoSet<RawVc>, EventListener>>;
    fn emit_collectible(
        &self,
        trait_type: TraitTypeId,
        collectible: RawVc,
        task: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn unemit_collectible(
        &self,
        trait_type: TraitTypeId,
        collectible: RawVc,
        task: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn update_task_cell(
        &self,
        task: TaskId,
        index: CellId,
        content: CellContent,
        turbo_tasks: &dyn TurboTasksBackendApi
    );
    fn get_or_create_persistent_task(
        &self,
        task_type: PersistentTaskType,
        parent_task: TaskId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> TaskId;
    fn create_transient_task(
        &self,
        task_type: TransientTaskType,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> TaskId;

    fn initialize(&mut self, task_id_provider: &dyn TaskIdProvider) { ... }
    fn startup(&self, turbo_tasks: &dyn TurboTasksBackendApi) { ... }
    fn stop(&self, turbo_tasks: &dyn TurboTasksBackendApi) { ... }
    fn idle_start(&self, turbo_tasks: &dyn TurboTasksBackendApi) { ... }
    fn try_read_own_task_cell_untracked(
        &self,
        current_task: TaskId,
        index: CellId,
        turbo_tasks: &dyn TurboTasksBackendApi
    ) -> Result<CellContent> { ... }

}

Required Associated Types

type ExecutionScopeFuture<T: Future<Output = Result<()>> + Send + 'static>: Future<Output = Result<()>> + Send + 'static

Required Methods

fn invalidate_task(&self, task: TaskId, turbo_tasks: &dyn TurboTasksBackendApi)

fn invalidate_tasks(
&self,
tasks: Vec<TaskId>,
turbo_tasks: &dyn TurboTasksBackendApi
)

fn get_task_description(&self, task: TaskId) -> String

fn execution_scope<T: Future<Output = Result<()>> + Send + 'static>(
&self,
task: TaskId,
future: T
) -> Self::ExecutionScopeFuture<T>

fn try_start_task_execution(
&self,
task: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Option<TaskExecutionSpec>

fn task_execution_result(
&self,
task: TaskId,
result: Result<Result<RawVc>, Option<Cow<'static, str>>>,
turbo_tasks: &dyn TurboTasksBackendApi
)

fn task_execution_completed(
&self,
task: TaskId,
duration: Duration,
instant: Instant,
stateful: bool,
turbo_tasks: &dyn TurboTasksBackendApi
) -> bool

fn run_backend_job<'a>(
&'a self,
id: BackendJobId,
turbo_tasks: &'a dyn TurboTasksBackendApi
) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>>

fn try_read_task_output(
&self,
task: TaskId,
reader: TaskId,
strongly_consistent: bool,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<Result<RawVc, EventListener>>

fn try_read_task_output_untracked(
&self,
task: TaskId,
strongly_consistent: bool,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<Result<RawVc, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn try_read_task_cell(
&self,
task: TaskId,
index: CellId,
reader: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<Result<CellContent, EventListener>>

fn try_read_task_cell_untracked(
&self,
task: TaskId,
index: CellId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<Result<CellContent, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn try_read_task_collectibles(
&self,
task: TaskId,
trait_id: TraitTypeId,
reader: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<Result<AutoSet<RawVc>, EventListener>>

fn emit_collectible(
&self,
trait_type: TraitTypeId,
collectible: RawVc,
task: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
)

fn unemit_collectible(
&self,
trait_type: TraitTypeId,
collectible: RawVc,
task: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
)

fn update_task_cell(
&self,
task: TaskId,
index: CellId,
content: CellContent,
turbo_tasks: &dyn TurboTasksBackendApi
)

fn get_or_create_persistent_task(
&self,
task_type: PersistentTaskType,
parent_task: TaskId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> TaskId

fn create_transient_task(
&self,
task_type: TransientTaskType,
turbo_tasks: &dyn TurboTasksBackendApi
) -> TaskId

Provided Methods

fn initialize(&mut self, task_id_provider: &dyn TaskIdProvider)

fn startup(&self, turbo_tasks: &dyn TurboTasksBackendApi)

fn stop(&self, turbo_tasks: &dyn TurboTasksBackendApi)

fn idle_start(&self, turbo_tasks: &dyn TurboTasksBackendApi)

fn try_read_own_task_cell_untracked(
&self,
current_task: TaskId,
index: CellId,
turbo_tasks: &dyn TurboTasksBackendApi
) -> Result<CellContent>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.