Trait turbo_tasks::graph::GraphStore

pub trait GraphStore<Node>: Default {
    type Handle: Clone;

    fn insert(
        &mut self,
        from_handle: Option<Self::Handle>,
        node: Node
    ) -> Option<(Self::Handle, &Node)>;
}

A graph store is a data structure that will be built up during a graph traversal. It is used to store the results of the traversal.

Required Associated Types

type Handle: Clone

Required Methods

fn insert(
&mut self,
from_handle: Option<Self::Handle>,
node: Node
) -> Option<(Self::Handle, &Node)>

Inserts a node into the graph store, and returns a handle to it.

If this method returns None, the node edges will not be visited.