Trait turbo_tasks::graph::GraphStore

pub trait GraphStore {
    type Node;
    type Handle: Clone;

    // Required method
    fn insert(
        &mut self,
        from_handle: Option<Self::Handle>,
        node: GraphNode<Self::Node>
    ) -> Option<(Self::Handle, &Self::Node)>;
}

A graph store is a data structure that will be built up during a graph traversal. It is used to store the results of the traversal.

Required Associated Types

type Node

type Handle: Clone

Required Methods

fn insert( &mut self, from_handle: Option<Self::Handle>, node: GraphNode<Self::Node> ) -> Option<(Self::Handle, &Self::Node)>

Inserts a node into the graph store, and returns a handle to it.

If this method returns None, the node edges will not be visited.