Trait turbo_tasks::graph::Visit

pub trait Visit<Node, Abort = !, Impl = ()> {
    type Edge;
    type EdgesIntoIter: IntoIterator<Item = Self::Edge>;
    type EdgesFuture: Future<Output = Result<Self::EdgesIntoIter>>;

    // Required methods
    fn visit(&mut self, edge: Self::Edge) -> VisitControlFlow<Node, Abort>;
    fn edges(&mut self, node: &Node) -> Self::EdgesFuture;
}

A trait that allows a graph traversal to visit the edges of a node transitively.

Required Associated Types

type Edge

type EdgesIntoIter: IntoIterator<Item = Self::Edge>

type EdgesFuture: Future<Output = Result<Self::EdgesIntoIter>>

Required Methods

fn visit(&mut self, edge: Self::Edge) -> VisitControlFlow<Node, Abort>

Visits an edge to get to the neighbor node. Should return a VisitControlFlow that indicates whether to:

continue visiting the neighbor node edges;

skip visiting the neighbor node’s edges;

abort the traversal entirely.

fn edges(&mut self, node: &Node) -> Self::EdgesFuture

Returns a future that resolves to the outgoing edges of the given node.