Struct turbo_tasks::primitives::Regex

#[repr(transparent)]
pub struct Regex(pub Regex);

Tuple Fields
0: Regex
Implementations

impl Regex

pub fn cell(self) -> RegexVc

Places a value in a cell of the current task.

Cell is selected based on the value type and call order of cell.