Trait turbo_tasks::InvalidationReason

pub trait InvalidationReason: DynamicEqHash + Display + Send + Sync + 'static {
    // Provided method
    fn kind(&self) -> Option<StaticOrArc<dyn InvalidationReasonKind>> { ... }
}

A user-facing reason why a task was invalidated. This should only be used for invalidation that were triggered by the user.

Reasons are deduplicated, so this need to implement Eq and Hash

Provided Methods

fn kind(&self) -> Option<StaticOrArc<dyn InvalidationReasonKind>>