Trait turbo_tasks::TryJoinIterExt

pub trait TryJoinIterExt<T, F>: Iteratorwhere
    T: Unpin,
    F: Future<Output = Result<T>>,{
    fn try_join(self) -> TryJoin<F>;
}

Required Methods

fn try_join(self) -> TryJoin<F>

Returns a future that resolves to a vector of the outputs of the futures in the iterator, or to an error if one of the futures fail.

Unlike Futures::future::try_join_all, this returns the Error that occurs first in the list of futures, not the first to fail in time.