Trait turbo_tasks::TurboTasksApi

pub trait TurboTasksApi: TurboTasksCallApi + Sync + Send {

    // Required methods
    fn invalidate(&self, task: TaskId);
    fn invalidate_with_reason(
        &self,
        task: TaskId,
        reason: StaticOrArc<dyn InvalidationReason>
    );
    fn notify_scheduled_tasks(&self);
    fn try_read_task_output(
        &self,
        task: TaskId,
        strongly_consistent: bool
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_output_untracked(
        &self,
        task: TaskId,
        strongly_consistent: bool
    ) -> Result<Result<RawVc, EventListener>>;
    fn try_read_task_cell(
        &self,
        task: TaskId,
        index: CellId
    ) -> Result<Result<CellContent, EventListener>>;
    fn try_read_task_cell_untracked(
        &self,
        task: TaskId,
        index: CellId
    ) -> Result<Result<CellContent, EventListener>>;
    fn read_task_collectibles(
        &self,
        task: TaskId,
        trait_id: TraitTypeId
    ) -> RawVcSetVc;
    fn emit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc);
    fn unemit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc);
    fn unemit_collectibles(
        &self,
        trait_type: TraitTypeId,
        collectibles: &AutoSet<RawVc>
    );
    fn try_read_own_task_cell_untracked(
        &self,
        current_task: TaskId,
        index: CellId
    ) -> Result<CellContent>;
    fn read_own_task_cell(
        &self,
        task: TaskId,
        index: CellId
    ) -> Result<CellContent>;
    fn update_own_task_cell(
        &self,
        task: TaskId,
        index: CellId,
        content: CellContent
    );
    fn mark_own_task_as_finished(&self, task: TaskId);
    fn connect_task(&self, task: TaskId);
    fn detached(
        &self,
        f: Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>>;

}

Required Methods

fn invalidate(&self, task: TaskId)

fn invalidate_with_reason( &self, task: TaskId, reason: StaticOrArc<dyn InvalidationReason> )

fn notify_scheduled_tasks(&self)

Eagerly notifies all tasks that were scheduled for notifications via schedule_notify_tasks_set()

fn try_read_task_output( &self, task: TaskId, strongly_consistent: bool ) -> Result<Result<RawVc, EventListener>>

fn try_read_task_output_untracked( &self, task: TaskId, strongly_consistent: bool ) -> Result<Result<RawVc, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn try_read_task_cell( &self, task: TaskId, index: CellId ) -> Result<Result<CellContent, EventListener>>

fn try_read_task_cell_untracked( &self, task: TaskId, index: CellId ) -> Result<Result<CellContent, EventListener>>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn read_task_collectibles( &self, task: TaskId, trait_id: TraitTypeId ) -> RawVcSetVc

fn emit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc)

fn unemit_collectible(&self, trait_type: TraitTypeId, collectible: RawVc)

fn unemit_collectibles( &self, trait_type: TraitTypeId, collectibles: &AutoSet<RawVc> )

fn try_read_own_task_cell_untracked( &self, current_task: TaskId, index: CellId ) -> Result<CellContent>

INVALIDATION: Be careful with this, it will not track dependencies, so using it could break cache invalidation.

fn read_own_task_cell(&self, task: TaskId, index: CellId) -> Result<CellContent>

fn update_own_task_cell(&self, task: TaskId, index: CellId, content: CellContent)

fn mark_own_task_as_finished(&self, task: TaskId)

fn connect_task(&self, task: TaskId)

fn detached( &self, f: Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>>

Wraps the given future in the current task.